s3gof3r: deal with errors in CompleteMultipartUpload
S3 will return a 200 but with an Error XML element instead of a 500 in case of
an internal error where it wants us to try again.

We'll either get the ETag field filled, or the Code one. If the latter, we know
we hit an error and try again if it's "InternalError"; otherwise we return an
error.