s3gof3r: handle being asked to delete an empty list of keys
We sometimes go through the same prefixes twice and find out that we've already
deleted the objects. In that case, we want to return an empty result (which does
contain the empty list of "all objects" we should  delete) instead of trying to
send the request to S3 as this fails the schema validation.