putter: retry the "close" of an object on EOF
This request is sent to ask S3 to create an object out of the multiple parts
we've uploaded. It can take some time to respond and we've seen issues when
connecting through a HTTP proxy where the connection will get cut.

Retry in this case just like if we'd received a 500 as the operation did not
fail, we just didn't see it succeed yet.