ObjectLister.Next(): never reset `err` to `nil`
First of all, there is a race. If `retryListObjects()` ever returns an
error, we set `l.err` and cancel the workers without sending anything
to `l.resultCh`. But if the other workers get canceled before `Next()`
is called, then it's possible for `Next()`'s check of `l.err` [1] to
not see the error, but then for `resultCh` to get closed before the
`select` sees it. In this case, the error would get lost entirely.

So what should we do instead? Currently, if `retryListObjects()` ever
returns an error, we set `l.err` and cancel the workers. Thus the
intent is for the `ObjectLister` instance to fail irrevocably as soon
as the first error occurs (e.g., it won't continue to return results
for other prefixes). To preserve those semantics, change
`ObjectLister.Next()` to leave `err` unchanged, thus continuing to
pass through the error if one has been detected.

[1] Never mind that the check is a data race itself; that will be
    fixed in a future commit.