Fix error handling data race
Previously, `l.currentErr` was set in a goroutine, but then read in
`Next()` with no happens-before relationship. This is a data race, and
could prevent the caller from learning about an error.

Instead, let the `errorgroup.Group` collect any errors from `worker()`
goroutines. When all goroutines have finished, write that error to a
new `finalErr` field, before closing `resultCh`.

Change `Next()` to copy `finalErr` to `currentErr` if it discovers
that `resultCh` has been closed.

Derive a context from the errgroup and pass it to the workers, so that
an error from one worker causes the others to terminate. This prevents
a deadlock waiting for other goroutines to finish.

This means that `Next()` no longer has to watch the context, because
it will proceed automatically when `resultCh` is closed. So remove the
`ctx` member from `ObjectLister`.