checkClose(): remove function (it is now unused)
This function was broken by

    4ecf3c1 (checkClose: check for nil io.Closer, 2015-07-20)

when it was changed to take its `err` parameter as a value rather than
as a pointer. This meant that its `err = cerr` only had a local
effect, and _didn't_ have the desired effect of causing its caller to
return an error. Therefore, many `Close()` errors would have been
lost.

The error was compounded by the many callers who used

    defer checkClose(resp.Body, err)

but where `err` wasn't a named return value. Therefore, even if the
caller had been changed to

    defer checkClose(resp.Body, &err)

, the deferred function invocation would not have had the desired
effect, because changing the local `err` variable would not have
changed the caller's return value anyway.

The commits leading up to this have fixed these errors at the callers
and removed users of `checkClose()`, changing the code to do explicit
error handling. Now we can remove the `checkClose()` function
entirely.