TestBP(): "fix" test
The reason for the test failure was that the old buffers were being
discarded, but at the very next iteration of the loop a new buffer had
to be allocated. But during the test, that buffer was also expiring
and being discarded, forcing yet another buffer to be allocated.

This pointless allocation/GC churn could also theoretically happen
during actual use of the library, though the default timeout is long
enough that it was probably not a serious problem.

Fix the problem by never discarding the front queue element if it is
the only element in the queue.

There are still some data races to be fixed.