Client.retryRequest(): fix an error-handling bug
The code for rewinding the body was overwriting the `err` from the
last request attempt. This can cause a failure: if all `c.nTry`
attempts fail, then the `Seek()` succeeds, then the method will return
`nil, nil`, which is invalid and causes the caller to dereference nil.
(This problem was observed once while running the `TestMulti` test.)

Fix this by (a) getting rid of the named return values, which obscure
things, and (b) check for `c.nTry` inside the loop rather than relying
on the loop falling though with the correct `err` value.