putter.wg: change what the `WaitGroup` counts
Previously, the wait group was incremented when `flush()` was about to
put a part into `putter.ch`, and decremented when `retryPutPart()`
exited after having dispensed with one part. This is unconventional.
Instead, let's count the number of workers that are still alive. Once
the channel is closed, the workers will all exit, which is an event
that we can wait for instead. This will make it easier to allow early
cancellation of the workers if there is an error without having to
worry about the part bookkeeping going awry.