putter: rationalize the management of goroutines
The old code had a few races, especially in the way that it accessed
`p.err` from multiple goroutines without synchronization. It also had
to do quite a bit of bookkeeping to split `Write()` data into
buffers (and crashy if more than one `Write()` call is made
concurrently).

So change the dataflow and how goroutines are managed:

* Get rid of `putter.err`. Instead, use an `errgroup.Group` to start
  the worker goroutines and collect their errors.

* Instead of `Write()` splitting the data into buffers by hand, use an
  `io.Pipe`. `Write()` writes into its write end, and a new worker,
  `queueParts()` reads from its read end using `io.ReadFull()` and
  passes the parts along to the workers.

* On errors, cancel the context and close the write end of the pipe.
  This (relatively) promptly unblocks all of the workers and allows
  them to clean up after themselves.

* Add a long comment how this all works together.

This change required test `TestPutterAfterError` to be temporarily
disabled, because it relied on being able to write directly to
`putter.err` (which, of course, was also racy).