Add a function to get status codes from RespError.
Some external callers want to be able to extract the HTTP status code
from RespErrors, so provide a function to do that.

Previously RespError was in the s3gopher package so it was exported,
but now that RespError exists as part of the internal s3client package,
it can't be imported outside this project.